Fix snapshots that expand the top/left of the screen
In the brigade codebase, we have a component that renders 1px outside
the screen. The recent code we've added to improve the detection of the
box wasn't taking into account that elements might render off-screen.

Adding a simple check to see if either top or left is negative fixes the
issue. I didn't address this for elements that expand right or bottom of
the screen as I think the screenshotting process already handles those
(by including the content).