Make DOM inspection easier when debugging
We're adding a `pointer-events: none` to the <body> element during a
happo run to prevent spurious things from happening because your mouse
cursor happened to be hovering the browser. This isn't great when using
`happo debug`, because it makes it harder to peek in to the component's
DOM. I'm adding a query param that can be used to turn it off, and then
we use that while in debug mode.