Merge review and diffs screen into one
Since we no longer have the approve/reject buttons on the review page,
it has essentially become the same one as the diffs page that we upload
to s3. By merging the two we avoid having to duplicate efforts when
making the page better.

There is one major difference between the old and the new review page.
Before, all files would be listed here, even those that don't have
diffs. Now, only those that are either new since the last run, or ones
that diff will be listed (same as for the uploaded diff page).