Improve header styling for diffs + debug pages
Making this stand out more from the rest of the content helps visually
structure the page.

I also moved the "Generated: <timestamp>" text to the top right corner
so that it wasn't in the way as much. This bit of information is
secondary anyway.

To make the page continue to look good when viewed on a small screen, I
made the flex items wrap.