Add review-demo.rb script
We want to vastly improve the diffs page, but that is slow going because
we don't have a super easy way to preview our changes. This commit fixes
this by adding a review-demo.rb script that starts up a server with some
dummy data in it so we can see what a diffs page might look like. This
should make it easier to improve this page.

Fixes #103