Make each diff or new example linkable
There are sometimes a lot of diffs on a page and sometimes you want
people to be able to quickly jump to just one of the diffs. If we make
each diff linkable, and turn the title into a link, we can enable this
behavior pretty easily.

This is currently unstyled, so we are getting default link colors here,
which isn't the prettiest thing in the world, but it works.

I'm also not really sure how well this will work for multibyte strings.

This will also allow us to add a table of contents at the top if we
want, or other things like that.