Add link icon to links on diff page
I wanted to change the color so that the links aren't using the default
link-blue. Before I do that, having an indicator that the header is
linkable will help. I went with a link icon from the material-design
icon library. It's base64 inlined so that we don't have external
dependencies (which is important for the uploaded/serialized diff page).