Use React to render Diffs page
We're about to make the diff review experience better. To make future
enhancements easier, we're moving over the UI to React. This will allow
us to work inside an environment that we're all (Galooshi members)
familiar with.

I've been pushing back on introducing unnecessary complexity into this
project. We've had a fruitful discussion in #61 about possible ways to
move forward, and what I'm presenting here is basically a middle ground
between what @lencioni is suggesting and what I've been advocating for.

I'm making the page depend on three externally hosted scripts:
- React - so that we can write nice UIs
- ReactDOM - so that we can hook up those UIs to the DOM
- babel - so that we can use jsx, string templates, arrow functions and
  such.

The in-browser transpilation done by babel isn't the fastest, but I
don't see this as a bottleneck at this point.

If the application continues to grow, it might be worth introducing a
build step to the process. But at this point I don't think it's
necessary.

When I was porting over the image_slug method, I had some issue
reimplementing it in javascript. So I decided to delegate to
window.btoa() instead, which will simply base64 encode the string. This
fixes a potential naming collision issue, but at the cost of making the
URL less readable. Simplicity and robustness won this time, but there's
a chance we'll move over to more readable slugs at some later point in
time.

This change shouldn't result in any differences in the UI (except for
the already-mentioned slug change). Future commits will focus on
improving it.