Break apart diff image
This commit is a step in the direction of improving the diffs page.
Instead of generating a big diff image consisting of the before, after
and overlay, we instead keep all images separated. After this change,
the before and after image will no longer have gaps in place where there
are additions/deletions. The overlayed image will still have that
though.

For now, the diffs UI is mostly kept the same as before. We're rendering
images left-to-right. Future commits will take advantage of the fact
that we now have individual images to work with.