Add "swiper" view
This view is inspired by the "Swipe" view from github.
https://github.com/blog/817-behold-image-view-modes

It has a cursor where the mouse was last seen, and clips the current
image to not display to the right of that point. For now, I'm making the
previous image a background-image. This made implementation pretty
straightforward, but it has the downside of not stretching the container
to the full height of the image. So if the previous image is taller than
the after image, you'll never see that part. I think that's actually
fine in this case.