Use constants for views types instead of raw strings
This makes the code easier to understand and more robust.

At one point I wanted to use Object.values, but that was not available
in my browser since we aren't polyfilling it, so I resorted to mapping
over Object.keys to get the values first. Not super great, but not too
bad either.