Make all propTypes required
Although we aren't even using propType checks at runtime since we
don't have a development build, it is nice for them to be as accurate
as possible for documentation purposes and for when we eventually
become sophisticated enough to have a build step. All of these props
are required on all of these components, so let's mark them as such.