Avoid positional key props
If the order of the array changes for some reason, these props won't
be working correctly. Since we expect image.current to be unique here
and to always exist in the image object, we can just use that.

While making this change, I noticed that the imageShape was incorrect.
The diff image will not always be present, but the current image will
be.