On hover, show what's before and what's after
I asked my team about how they would make happo diffs better. One
interesting suggestion was to include a label for which image is the
before and after. It's easy to forget which one is the old and new one I
guess.

I'm making the label appear on hover only. It's located in the lower
right corner.