Memoize directory in uploader
In 602f49a7 we extracted this method from just a one-time variable
setting. This broke image uploads because it generates a new random
directory every time it is called, which causes all of the images and
the uploaded HTML to be in different spots.

We can fix this by memoizing it.

Fixes #129.