Add better spec for disabling scrollbars
In 239611300f1 I took a swing at hiding scrollbars, but in practice this
seemed to have no effect. I have a hypotheses that we aren't giving the
browser a chance to repaint the element before taking the screenshot,
but I wanted to write a test that verifies this first. I believe that
this might be such a test.

The idea is that if there is no scrollbar, the entire element will be
red. If there is a scrollbar, this will not be true because the
scrollbar itself is not entirely red.

Locally this test passes, which is consistent with the failure mode for
my previous attempt. If this test fails in CI, then we actually have
something to work with here. Update: this test passes in CI, so we are
back to square one. I think this is still a good test to have, so I'm
adding it anyway.

While working on this I discovered that the screenshot that we take of
elements includes the size of nodes that are partly hidden by overflow
containers and opened an issue at #136.