Start switch to in-browser lcs diffing
This is the first step in the direction of replacing the time-consuming
lcs-diffing that we do on the server to the client. The idea is that we
create a (lcs powered) diff dynamically on the client when you view the
diff page. The server can then use a much simpler (and faster)
equality-based diffing algorithm.

I found jsondiffpatch to help out with the diffing. I don't know if it's
good enough, but we'll give it a try. One big plus was that it came with
a browser-ready bundle so that we can depend on it directly through
unpkg.com.

https://github.com/benjamine/jsondiffpatch