Add method to get image data in <LCSDiff>
We are working to render the diff image on the client. One important
step with this is to be able to get the image data so we can actually
compute a diff. This commit does this by rendering the images into
canvas elements, collecting the data, and breaking it up into a 2D array
for convenience when diffing. We haven't tested this very well, so it is
likely that we have an off-by-1 error or other similar problems in here.