Draw simple diff image to canvas
This commit takes another step in the transition to a full in-browser
diff image renderer. We're not using any LCS diffing yet, and instead of
the magenta faded pixel we're using a bright red.

I'm not 100% sure I got the euclideanDistance function right. It looks
right-ish when in use, but I mosly copied this stuff from a SO thread:
http://stackoverflow.com/questions/13586999/color-difference-similarity-between-two-values-with-js

To prevent creating unnecessary arrays I created a `getPixelAt` method
that will grab the right rgba values from the one-dimensional array used
internally by the canvas.