Use webpack to bundle Happo
This only applies to the Happo client used when looking at diffs. In the
future we might want to switch over happo-runner.js as well.

To make developing easier, I added a `npm run webpack-dev` script that
will watch js file changes and rebuild the bundle on change. That script
will also produce a source map which is useful for debugging.

The "production" bundle (the one bundled when publishing the gem) will
be constructed using `npm run webpack` and will produce a minified
bundle without source maps. That bundle is currently at 14.8 KB.

I decided to continue to depend on an externally loaded React version.
We want to keep the bundle small, because it is injected into every diff
page as a text blob.

For now, the folder structure (under js/src) is flat. We might want to
restructure that in the future. I also kept all component/module names,
which don't always make sense. We could make a sweep later to improve
naming.