Improve scrollbar-prevention
After upgrading to 2.7.3 (from 2.5), we're seeing a bunch of mysterious
diffs coming from the code that is meant to prevent scrollbars.

[Fixes #146]
[Fixes #145]

Instead of applying `overflow: hidden` whenever we notice something
overflowing, we now also check to see that there's an `overflow: auto`
or `overflow: scroll` present. This check is a bit expensive, but if we
can do it only in cases of actual content overflow we're mostly
protected from performance issues.