Provide way to control root nodes when snapshotting
The fact that we now iterate through all elements to get the bounding
box to screenshot means that it's much harder to control the width of
the resulting snapshot image. At Brigade, we use React to render, which
means we create a div as the react root, and then render our
component(s) in this root. Before, we'd just make sure that the examples
returned the inner nodes (direct child of react root), but now we can't
do that. Which means that it's the bounding rectangle of the react root
div that's used when snapshotting.

To make it easier to control what nodes are used to get the bounding
box, I'm adding a `getRootNodes` method. The default simply returns
`document.body.children`, but it allows Brigade to override with this:

```js
happo.getRootNodes = function() {
  return document.querySelectorAll('[data-reactroot]');
};
```

React will automatically add a `[data-reacroot]` to all elements
rendered with `ReactDOM.render`. So this is a general React-specific
override that I intend to add to the README shortly.