Don't assume elements start at top left corner
Due to how we initialized the box at the top left corner, all snapshots
always started at the top left. This isn't great if you have a component
rendering at the right or bottom edge of the screen.

In the Brigade codebase, we have a <Toast> component that renders at the
bottom of the screen. Before this change, the snapshot would include a
large gray box above the component. After this change, only the
component is included in the snapshot.