Annotate diff page with link back to source/commit
To better link the diff page with a code change, we now allow you to
pass an optional URL as an argument to `happo upload_diffs`. This URL
will be used on the diffs page as a link, next to the label "Triggered
by:".

e.g.

happo upload_diffs "http://example.com"