Use adiff to power <LCSDiff>
This has been long time coming, but I finally have something that uses
LCS and draws something on the screen. It's terribly slow atm, but I
have some ideas on how to improve that.

The basic approach is that we first construct 2d arrays of the images,
then use the `diff` method from the adiff package to compute the
differenes. Adiff uses LCS under the hood, so the results are similar to
what we've had in the past.

We then use the diff results to inject rows of pixels into the before
and after image. In places where there is an addition, we inject green
pixels. In places where there is a deletion, we inject red pixels. We
then iterate over these images and pixel-by-pixel constructs the diff
image representation. Right now you can't really see the red/green
pixels, because `getDiffPixel` isn't smart enough. But that should be
fairly trivial to fix.

The code is in a very rough state, and I expect to clean this up
significantly in subsequent commits.