Compute LCS diff on hashed rows
This is a major performance enhancement. Instead of computing the diff
on the original pixels, we first hash them into a base64 string. This is
okay since we only care about knowing whether rows changed, not the
actual changes themselves. We still use the original pixels to draw the
image diff.