Optimize euclideanDistance()
We potentially run this many times in a tight loop that we want to make
faster. Since we know that rgba1 and rgba2 are arrays with 4 elements
each, we can ditch the loop and improve performance.

I measured this in a small example using `performance.now()` and this
seems to improve the performance of `getDiffPixel` on average by about
20%.

This function is a good candidate for SIMD, once that lands.