Fix diffs for images of different size
The previous interpretation of the adiff format was making the
assumption that deletions and additions were never mixed in the same
result. By taking both into account, and adding only those rows that the
image doesn't already have, we end up with an algorithm that can handle
images of different size.