Use web worker to construct diff image
This might not be the path we end up going down. But I wanted to at
least get some experience of using worker threads to do some of the
heavy lifting. This has a few advantages:

- We allow the UI to be responsive while processing diffs
- We can parallelize stuff (not done yet)

I found the web-worker webpack loader that makes the job of creating a
worker a lot simpler.