Restructure <LCSDiff>
Instead of having to deal with undefined values in state, we can use
Promise.all to wait for both images to resolve until we compute the
diffs. This led to a simpler setup.

Promise.all will resolve promises in parallell, but the order of the
result array will be the same as the order on which promises were
invoked.