Use web worker to compute diff
We want to avoid doing expensive diffing work on the main thread.
Leaving the diffing work up to a web worker will make the UI more
responsive, and might help us speed things up (by parallelizing) in the
future.

Right now, this adds a significant performance decrease. This is because
we're serializing/deserializing large arrays back and forth between the
worker and the main thread. I'm looking into using ArrayBuffer (which is
a transferable object [1]) instead of straight up arrays.

[1]: http://stackoverflow.com/questions/19152772/how-to-pass-large-data-to-web-workers