Keep diff views mounted
The LCSDiff view takes a while to compute, so we don't want to redo that
every time you switch to that view. By using a wrapping component that
simply renders the child in an invisible div when it's not supposed to
be visible we make sure that the component isn't unmounted.

To prevent prematurely mounting a view, we only cache the results when
the component has been made visible the first time.