Don't compute LCSDiff until it is in view
Using react-waypoint, we can make sure that the LCSDiff isn't computed
until you're actually looking at it.

I think I found a bug with react-waypoint in that a side-scrolled
ancestor is picked up as scrollableAncestor even though it shouldn't. I
reported this in [1]. To work around this I specified that the window
should be used.

[1]: https://github.com/brigade/react-waypoint/issues/118