Set min-height for image containers
To prevent the page from jumping around as diffs are computed/images are
loaded, we now store a `height` for each entry in the
result_summary.yaml file. We then use that value to set a min-height on
the divs that wrap examples (both new and diffs). Because of how the
LCSDiff view might end up being taller than this height, I'm adding a
bit extra to the diff view min-height. This will allow the LCSDiff to
grow a bit without causing the page to jump.