Render diffs on white background
We've been tweaking the transparency of the rendered colors, which makes
them a little hard to see and interpret. This commit brings in what we
have been doing on the Ruby side by composing the colors on top of a
white background.

For this version, I did a straight up port of the functions we were
using from ChunkyPNG. The performance considerations here might be
different than they were in Rubyland, so we should look into any
opportunities to optimize this code. Depending on what we learn here,
there may also be some opportunities to optimize some other maths that
we are doing, such as in euclideanDistance.