Initialize canvas in <LCSDiff> at zero width and height
The default canvas size (in Chrome at least) is 300 by 150 pixels. On
initial render, the canvas would be of this size in <LCSDiff>. This
caused the page to be taller than necessary, and made waypoints work in
funky ways (didn't trigger on mount). By setting an initial width and
height of zero, we avoid the jumpiness.