Make progress indication more granular
We think that we will be able to produce a nice progress animation. To
make it as smooth as possible, we'll be using an exponential backoff.
I'm not adding that just yet, but I'm making sure that we report back a
more fine-grained progress from ComputeAndInjectDiffsWorker. I also
adjusted numbers a little to better reflect real progress.