Replace <progress> with <SmoothProgress>
The new progress bar uses exponential backoff to make it look like
progress is happening even though we might be stalled on adiffing for
instance.

I adjusted some progress numbers so that we reach higher numbers earlier
in the process. I also made sure that the width and height for the
LCSDiff component is set earlier. This makes the styling of the progress
bar appear less flash-y.