Actually use 2D arrays in ComputeAndInjectDiffsWorker
I found this file to be confusing because we were using 3D arrays where
we said we were using 2D arrays. Since we don't actually need 3D arrays,
I decided to simplify it a bit. In my testing, the perf seems to be
about the same, but I would expect the memory usage to be a bit better,
since we are using typed arrays.