Avoid flattening 2D images
We are working to add a gutter to the diff image, and realized that the
right place for this is ImageDiffWorker. In order to make this easier,
we wanted to be able to operate on 2D image data, so we started by
moving the image flattening into ImageDiffWorker. Once that was done, we
noticed that we don't need to flatten these images at all.

We were worried that this might affect the performance of worker
communication, but it seems to be about the same to maybe a little bit
better.