Add topOffset to waypoint
If you land on a diff page via an inline link partway down the page,
scrolling up is a bit of an awkward experience because the images don't
load until you scroll to the top of the image to trigger the waypoint.

We could solve this by adding a waypoint at the bottom of the image, but
for now I think simply adding a topOffset to the current waypoint will
work okay. To help account for the height of the image, I am setting it
a little higher than the bottomOffset.

This should work well for most cases, but won't work great for very tall
images. We might want to revisit this at some point.