Properly escape uploaded image names
As reported in #156, we weren't escaping image names containing plus
signs.

URI.escape('a + b') will produce
'a%20+%20b'
which is treated by the browser as 'a   b'.

We need something that escapes plus signs as well, just like
encodeURIComponent does on the javascript side of things. I'm surprised
there's nothing better than having to pass a regexp to URI.escape in
Ruby, but that seems to be how things are done [1].

We need tests around this, but I'll punt on that until we have things
converted to js.

[1] http://stackoverflow.com/questions/2834034/how-do-i-raw-url-encode-decode-in-javascript-and-ruby-to-get-the-same-values-in-b