Prevent runaway loops in longestCommonSubsequence
When computing the longest common subsequence, the inner loop could end
up doing much more work than is likely to be useful. To prevent this
from happening, we are limiting it to 100 places in either direction.
This should work well for image diffs without sacrificing diff quality
too much.

In our non-scientific tests, we noticed that this change reduced the
time in this function from ~700ms to ~200ms for a largish image. The
performance savings will vary greatly based on a case-by-case basis, so
this savings is somewhat of a worst case scenario.

We might want to make this configurable at some point, e.g. when we move
this to its own package.