Change hashing from JSON.stringify back to btoa
When we switched from btoa to JSON.stringify, the latter outperformed
the former. Now that we've rewritten our LCS implementation,
inexplicably the JSON.stringify version is sometimes about 5 times
slower than the btoa version. We don't really understand why, but this
seems to be pretty consistent (at least in Chrome). Changing this back
brings the hashing of the globe images down from 2552ms to 682ms.