Fix image hashing for Safari
Apparently there is a bug in Safari where trying to reference `btoa`
inside a web worker will result in an error, so we fall back to the
slower (?) `JSON.stringify`. The only way to prevent this seems to be by
using a try/catch. We do this in its own function to prevent our align
function from being de-optimized.

https://bugs.webkit.org/show_bug.cgi?id=158576