Use JSON.stringify instead of btoa in Firefox
Firefox, for some reason, gives us the same string when feeding typed
arrays to `btoa`. Here, we can detect this behavior and fall back to the
slower (?) but more accurate `JSON.stringify`. Since we are using
feature detection for this, this should make it safer in any browser and
also so it will automatically upgrade whenever this is fixed.

According to MDN, btoa() works on strings and if you want to work on
ArrayBuffers, we need to write our own implementation. [0] We looked at
using this array buffer to base64 method

  https://gist.github.com/jonleighton/958841

but it wasn't great for performance, so we decided against it and to
roll with JSON.stringify instead.

[0]: https://developer.mozilla.org/en-US/docs/Web/API/WindowBase64/Base64_encoding_and_decoding#Solution_.232_.E2.80.93_rewriting_atob()_and_btoa()_using_TypedArrays_and_UTF-8