Clean out old worker files when building gem
I looked in the latest gem that we released and found a bunch of these
old worker files built up that we don't actually need to ship. To
prevent this from happening going forward, I added an `npm run clean`
script that will remove them and an `npm run build` scrip that first
cleans and then runs webpack.