Avoid namespacing functions that don't need to be
When HappoRunner.js was still a non-transpiled, raw javascript file, we
needed functions to be namespaced under `happo` in order to avoid
polluting the global namespace. Now that we're in webpack land, where
modules are automatically given a local scope, we don't have to do that
anymore. I've moved some functions to the top of the file, and I've
extracted one external module (removeScrollbars.js).