Don't run npm stuff in gemspec
When we switched over to Babel and Webpack to compile js, I added two
lines at the top of the gemspec file to make sure that the js was
compiled as part of building a new gem version (`gem build
happo.gemspec`). This ended up making tests a lot slower, because we
would kick of a full npm install and webpack build in a number of places
during the Travis run. For instance, the gemspec file is read when doing
`bundle exec`.

Instead of automatically compiling the bundles, I'm adding a big warning
at the top of the gemspec file to remind myself (and others) that they
need to compile themselves as part of releasing a new version.

I believe this should make Travis builds a lot faster.