Add `happojs` command-line tool
This command will eventually be renamed to just `happo`, but while still
developing on it, I wanted to avoid naming collisions locally.

For now, the only supported commands are `debug` and `review-demo`.
These both just start the server and give you a URL to open. We might
want to merge these under a common name at some point, but for now I'm
keeping them separate.

I've used commander [1] to power the cli tool. We have some experience
from it in the past when working on import-js, and I think it's a good
fit for this project as well.