Add first step towards running diffs in js
This commit is a POC that we can run selenium in javascript and bring up
a web page where happo is defined. We'll continue building on top of
this framework.

When you run `happojs run` now, a firefox instance is opened. We then
navigate to the index page for happo (where diffing will happen). To
verify that the page is working, I've added a check for `window.happo`
being defined.

When I first installed selenium, it gave me 3.0-beta. That version
requires geckodriver to power firefox, but that doesn't fully work yet
as far as I know. So I installed the latest stable version, 2.53,
instead.