Add rendering and screenshotting to JS version
There's still work to be done here, where we left some TODOs, but this
commit brings us on the next step of our journey of rewriting in
JavaScript.

Due to the promise-based interface, we ended up doing a recursive
solution again. Although, I wonder if we can get more speed by making
this asynchronous starting at the cropping and using Promise.all to
process all of the screenshots at the same time.