Start saving PNGs to snapshots folder
This is still work-in-progress, but it takes us a step closer to having
a functional happo runner in javascript.

We're using pngjs (which png-crop depends on) to read and write images
to disk. I've ported over code from the old runner.rb file.

After this change, snapshots will be saved to disk. There's an
unresolved issue where an error is thrown on the second happojs run:

  Error: There are some read requests waitng on finished stream

(spelled incorrectly because that's what is logged)

Apart from fixing that error, we just need to construct a result summary
file. And then we're done with the runner.