Use cssesc to power imageSlug.js
To be able to link directly to a diff on the diff page, we use `id`
attributes to create a unique location on the page. Before, we used
`btoa` to escape a string that could be used in an `id` attribute. But
btoa can only handle latin characters so when you used something like
curly quotes (), the diffs page would break. I changed us over to
`cssesc` [1] instead. In the future, we can probably rely on
CSS.escape(), but until that has wide enough adoption the library will
do just fine.

Fixes [#164]

[1]: https://github.com/mathiasbynens/cssesc