Move index route over to /snapshot
I want to improve a few defaults for happo. For instance, running just
`happo` shouldn't do anything else than output help information. This is
because it's more common to use `happo debug` while developing, and I
make the mistake of running `happo` all the time. At some point, we
might want to switch over the default to just be debug, but for now I
want to make it explicit.

This commit doesn't actually implement ^^, but instead just renames some
routes and files so that there's no index (`/`) route anymore. It's now
`/snapshot`.