Output help if no arg provided to cli
I thought I would have to remove the default behavior of running the
full diffing process when no argument was provided to the cli tool. But
it turns out we never ported this default behavior over from the ruby
cli. So instead I'm just adding some help output if you forget to
provide a command.