Add basic test for getFullRect()
We want to start writing unit tests to ensure robustness and good test
speed. This module seems like a logical place to start.

We ran into an issue with jsdom returning all 0 values from
getBoundingClientRect() so we had to mock it out. This will probably be
the case until jsdom gets a layout engine.

  https://github.com/tmpvar/jsdom/issues/1322

Our current createElement function is pretty rudimentary and will likely
need to be adjusted as we flesh out our tests here, particularly with
nested examples.

Likewise, we also discovered a difference in how jsdom treats
getComputedStyle. In the browser, all valid properties are represented
but in jsdom, only the properties that have been explicitly set are
represented. This means that getting the property value of a computed
style that was not explicitly set (e.g. margin in our example) will
return '0px' in a browser, but '' in jsdom. Since we pass this value
straight into parseFloat, we were seeing NaN instead of 0. We fixed this
by having our getPropertyValue calls fall back to 0.